requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dj-config-url==0.1.1
dj-database-url==1.2.0
Django==4.1.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
payme-pkg==2.1
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.2.0
